+++ keystatic.config.ts
@@ -2,8 +2,10 @@ import { config, fields, collection, singleton } from "@keystatic/core";
 
 export default config({
   storage: {
-    kind: "github",
-    repo: "jesusdavidweb/eleahora",
+    kind: "cloud",
+  },
+  cloud: {
+    project: "eleahora/eleahora",
   },
 
   ui: {